pyproject.toml
[tool.poetry]
name = "automate_mappings"
version = "0.1.0"
description = "Uses linkml/schema-automator to suggest mappings to biolink classes in external ontologies"
authors = ["Sierra Taylor Moxon <sierra.taylor@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
schema-automator = "^0.2.8"
biolink-model = "^3.0.2"
tox = "^3.25.1"
pytest = "^7.1.3"
oaklib = "^0.1.45"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
